feat: 互动弹幕/视频分段进度条/互动弹幕显示初步

+++ app/src/main/java/cn/spacexc/wearbili/remake/app/video/info/comment/domain/CommentRemoteDto.kt
@@ -57,7 +57,7 @@ data class CommentContentData(
     var rcount: Int,
     var state: Int,
     var fansgrade: Int,
-    var attr: Int,
+    var attr: Long,
     var ctime: Long,
     var rpid_str: String,
     var root_str: String,
